requirements.txt
streamlit~=1.36.0
pillow~=10.4.0
propbot~=0.0.1
python-dotenv~=1.0.1
langchain-community~=0.2.16
langchain~=0.2.16
matplotlib~=3.8.1
llama-index-core~=0.11.8
pathlib~=1.0.1
docx~=0.2.4
python-docx~=1.1.2
fitz~=0.0.1.dev2
camelot-py~=0.11.0
pypdf2~=3.0.1
pandas~=2.2.2
qdrant-client~=1.11.1

- e .